Crypten hook logic shifted to hook_crypten.py (#3406)
* crypten hook logic shifted to crypten/hook/hook.py

* Add a missing `NodeClient` import to `AbstractGrid`
